+++ account/templates/account/index.html
@@ -1,0 +1,4 @@
+{% extends 'index.html' %}
+{% load static %}
+{% block content %}
+{% endblock %}
