+++ account/templates/login.html
@@ -3,7 +3,7 @@
 {% block content %}
   <div class = "login_field" style="text-align: center; margin: 100px">
     <h2>Sign In</h2>
-    <form action= {% url 'login' %} method="post">
+    <form action= "{% url 'login' %}" method="post">
       {% csrf_token %}
       {{ form.as_p }}
       <button type="submit">Sign In</button>
