+++ account/urls.py
@@ -1,6 +1,9 @@
 from django.contrib import admin
 from django.urls import path, include
+from . import views
 
 app_name = 'account'
 
-urlpatterns = []
+urlpatterns = [
+    path('', views.index, name='index'),
+]
