+++ courses/views.py
@@ -1,17 +1,15 @@
+from django.contrib.auth.models import User
+from .forms import CSVUploadForm
+from django.shortcuts import render, redirect
+from django.http import JsonResponse, HttpResponse, HttpResponseRedirect
+from django.views.decorators.csrf import csrf_exempt
+from .models import Course, Tag
+from rest_framework import permissions
+from rest_framework.decorators import api_view, permission_classes
 import csv
 from io import TextIOWrapper
-from django.contrib.auth.models import User
-
-from django.shortcuts import render, redirect
-
-from courses.forms import CSVUploadForm
-from .models import Course, Tag
-
-from django.middleware.csrf import get_token
-from django.urls import reverse
 from datetime import datetime
 from decimal import Decimal, InvalidOperation
-from django.forms.models import model_to_dict
 
 
 def total_course(request):
@@ -101,3 +99,84 @@ def upload_csv(request):
     else:
         form = CSVUploadForm()
     return render(request, 'admin/upload.html', {'form': form})
+
+
+@csrf_exempt
+@api_view(['POST'])
+@permission_classes([permissions.IsAuthenticated])
+def like_course(request, course_id):
+    course = Course.objects.get(id=course_id)
+    user_id = request.data.get('user_id')
+    if user_id:
+        user = User.objects.get(id=user_id)
+        course.likes.add(user)
+        course.dislikes.remove(user)
+        return JsonResponse({}, status=201)
+    else:
+        return JsonResponse({'error': 'Invalid request.'}, status=400)
+
+
+@csrf_exempt
+@api_view(['POST'])
+@permission_classes([permissions.IsAuthenticated])
+def dislike_course(request, course_id):
+    course = Course.objects.get(id=course_id)
+    user_id = request.data.get('user_id')
+    if user_id:
+        user = User.objects.get(id=user_id)
+        course.dislikes.add(user)
+        course.likes.remove(user)
+        return JsonResponse({}, status=202)
+    else:
+        return JsonResponse({'error': 'Invalid request.'}, status=400)
+
+
+@csrf_exempt
+@api_view(['POST'])
+@permission_classes([permissions.IsAuthenticated])
+def check_course_like(request, course_id):
+    course = Course.objects.get(id=course_id)
+    user_id = request.data.get('user_id')
+    if user_id:
+        user = User.objects.get(id=user_id)
+        if user in course.likes.all():
+            check = 1
+        else:
+            check = 0
+        return JsonResponse({'check': check})
+    else:
+        return JsonResponse({'error': 'Invalid request.'}, status=400)
+
+
+# TODO 이유는 모르겠지만 빈 wishlist반환됨.
+
+
+# @csrf_exempt
+# @api_view(['POST'])
+# @permission_classes([permissions.IsAuthenticated])
+# def user_wishlist(request):
+#     user_id = request.data.get('user_id')
+#     if user_id:
+#         user = User.objects.get(id=user_id)
+#         interests = user.userprofile.interests.all()
+#         wishlist = []
+#         for interest in interests:
+#             wishlist.append({
+#                 'id': interest.id,
+#                 'courses': {
+#                     'course_id': interest.id,
+#                     'course_name': interest.title,
+#                 }
+#             })
+#         return JsonResponse({'wishlist': wishlist}, status=200)
+#     else:
+#         return JsonResponse({'error': 'Invalid request.'}, status=400)
+
+
+@csrf_exempt
+@api_view(['GET'])
+@permission_classes([permissions.IsAuthenticated])
+def course_like_count(request, course_id):
+    course = Course.objects.get(id=course_id)
+    count = course.likes.count()
+    return JsonResponse({'like_count': count})
